Collect all properties in betty.properties
