+++ betty/media_type/__init__.py
@@ -15,10 +15,9 @@ from betty.pathlib import StrPath
 from betty.plugin import PluginTypeDefinition
 from betty.plugin.ordered import Order, OrderedPluginDefinition
 from betty.portable import Portable, PortableData
-from betty.property import Property
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable, Mapping, Sequence
+    from collections.abc import Iterable, Mapping, Sequence
 
     from betty.locale.localizable import ResolvableLocalizable
     from betty.machine_name import ResolvableMachineName
@@ -224,32 +223,6 @@ class MediaTypeDefinition(HumanFacingDefinition, OrderedPluginDefinition):
 
 
 type ResolvableMediaType = MediaType | MediaTypeDefinition
-
-
-@final
-class MediaTypeProperty(Property[MediaType, ResolvableMediaType]):
-    """
-    A property containing a media type.
-    """
-
-    def __init__(
-        self,
-        *,
-        label: ResolvableLocalizable | None = None,
-        default: Callable[[], MediaType] | None = None,
-        description: ResolvableLocalizable | None = None,
-        omit_load: bool | None = None,
-        omit_dump: Callable[[MediaType], bool] | None = None,
-    ):
-        super().__init__(
-            data=MediaType,
-            default=default,
-            label=label,
-            description=description,
-            omit_load=omit_load,
-            omit_dump=omit_dump,
-            resolver=resolve_media_type,
-        )
 
 
 def resolve_media_type(media_type: ResolvableMediaType, /) -> MediaType:
