+++ betty/plugin/data/ordered.py
@@ -12,7 +12,7 @@ from betty.machine_name import MachineName
 from betty.plugin import PluginDefinition
 from betty.plugin.data import PluginDefinitionConfiguration
 from betty.plugin.resolve import ResolvablePluginId, resolve_plugin_id
-from betty.property.collection.sequence import SequenceProperty
+from betty.properties.collection.sequence import SequenceProperty
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
