+++ betty/plugins/content/box.py
@@ -13,10 +13,12 @@ from betty.data.aggregate.record.object import ObjectDefinition
 from betty.data.str import StrDefinition
 from betty.factory import DataManufacturable
 from betty.locale.localizable.gettext import _
-from betty.plugin.data.property import PluginManufacturerSequenceProperty
 from betty.plugins.content.render import Render, RenderConfiguration
 from betty.plugins.content.template import Template, TemplateBuild
 from betty.project import Project
+from betty.properties.plugin_manufacturer_sequence import (
+    PluginManufacturerSequenceProperty,
+)
 from betty.property import Optional, Property
 from betty.sample import Sample, Size
