+++ betty/plugins/content/raspberry_mint_color_style.py
@@ -13,11 +13,13 @@ from betty.data.aggregate.record.object import ObjectDefinition
 from betty.data.enum import EnumDefinition
 from betty.factory import DataManufacturable
 from betty.locale.localizable.gettext import _
-from betty.plugin.data.property import PluginManufacturerSequenceProperty
 from betty.plugins.asset_directory.raspberry_mint import RASPBERRY_MINT
 from betty.plugins.content.template import Template, TemplateBuild
 from betty.plugins.extension.raspberry_mint import ColorStyle as RaspberryMintColorStyle
 from betty.project import Project
+from betty.properties.plugin_manufacturer_sequence import (
+    PluginManufacturerSequenceProperty,
+)
 from betty.property import Property
 from betty.sample import Sample
