+++ betty/plugins/entity/note.py
@@ -15,9 +15,9 @@ from betty.entity.has_links import HasLinks
 from betty.entity.has_media_type import HasMediaType
 from betty.locale.localizable.gettext import _, ngettext
 from betty.locale.localizable.linked_data import dump_linked_data
-from betty.locale.localizable.property import LocalizableProperty
 from betty.locale.localizable.static.schema import StaticTranslationsSchema
 from betty.privacy import HasPrivacy, Privacy, is_public
+from betty.properties.localizable import LocalizableProperty
 
 if TYPE_CHECKING:
     from betty.entity.has_notes import HasNotes
