+++ betty/plugins/entity/source.py
@@ -20,9 +20,9 @@ from betty.entity.has_notes import HasNotes
 from betty.linked_data import JsonLdObject, dump_context
 from betty.locale.localizable.gettext import _, ngettext
 from betty.locale.localizable.linked_data import dump_linked_data
-from betty.locale.localizable.property import LocalizableProperty
 from betty.locale.localizable.static.schema import StaticTranslationsSchema
 from betty.privacy import HasPrivacy, Privacy, is_public, merge_privacies
+from betty.properties.localizable import LocalizableProperty
 from betty.property import Optional
 
 if TYPE_CHECKING:
