+++ betty/properties/__init__.py
@@ -1,0 +1,3 @@
+"""
+Property implementations.
+"""
