+++ betty/properties/collection/sequence.py
