+++ betty/properties/countable_localizable.py
@@ -1,5 +1,5 @@
 """
-Localizable attributes.
+Countable localizable properties.
 """
 
 from __future__ import annotations
@@ -8,37 +8,12 @@ from typing import final
 
 from betty.locale.localizable import (
     CountableLocalizable,
-    Localizable,
     ResolvableCountableLocalizable,
     ResolvableLocalizable,
     resolve_countable_localizable,
-    resolve_localizable,
 )
-from betty.locale.localizable.data import (
-    CountableLocalizableDefinition,
-    LocalizableDefinition,
-)
+from betty.locale.localizable.data import CountableLocalizableDefinition
 from betty.property import Property
-
-
-@final
-class LocalizableProperty(Property[Localizable, ResolvableLocalizable]):
-    """
-    A property containing a :py:class:`betty.locale.localizable.Localizable`.
-    """
-
-    def __init__(
-        self,
-        *,
-        label: ResolvableLocalizable,
-        description: ResolvableLocalizable | None = None,
-    ):
-        super().__init__(
-            LocalizableDefinition(),
-            label=label,
-            description=description,
-            resolver=resolve_localizable,
-        )
 
 
 @final
