+++ betty/tests/locale/test_data.py
@@ -1,5 +1,5 @@
 from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG
-from betty.locale.data import LocaleDefinition, LocaleProperty
+from betty.locale.data import LocaleDefinition
 
 
 class TestLocaleDefinition:
@@ -10,13 +10,3 @@ class TestLocaleDefinition:
     def test_dump(self) -> None:
         sut = LocaleDefinition()
         assert sut.porter.dump(DEFAULT_LOCALE) == DEFAULT_LOCALE_TAG
-
-
-class TestLocaleProperty:
-    class _Owner:
-        locale = LocaleProperty()
-
-    def test_set(self) -> None:
-        sut = self._Owner()
-        sut.locale = DEFAULT_LOCALE_TAG
-        assert sut.locale == DEFAULT_LOCALE
