+++ betty/tests/properties/collection/test_sequence.py
@@ -2,7 +2,7 @@ from betty.data import Data
 from betty.data.aggregate.collection.sequence import SequenceDefinition
 from betty.data.aggregate.record.object import ObjectDefinition
 from betty.data.str import StrDefinition
-from betty.property.collection.sequence import SequenceProperty
+from betty.properties.collection.sequence import SequenceProperty
 from betty.test_utils.locale.localizable import DUMMY_LOCALIZABLE
