+++ betty/tests/properties/test_localizable.py
@@ -1,12 +1,9 @@
 from gettext import NullTranslations
 
-from betty.locale import DEFAULT_LOCALE, DEFAULT_LOCALE_TAG
+from betty.locale import DEFAULT_LOCALE_TAG
 from betty.locale.localizable.plain import Plain
-from betty.locale.localizable.property import (
-    CountableLocalizableProperty,
-    LocalizableProperty,
-)
 from betty.locale.localize import DEFAULT_LOCALIZER, Localizer
+from betty.properties.localizable import LocalizableProperty
 from betty.test_utils.locale.localizable import DUMMY_LOCALIZABLE
 
 
@@ -37,30 +34,3 @@ class TestLocalizableProperty:
         localizable = Plain("Hello, world!")
         instance.attr = localizable
         assert instance.attr is localizable
-
-
-class TestCountableLocalizableProperty:
-    class _Instance:
-        attr = CountableLocalizableProperty(label=DUMMY_LOCALIZABLE)
-
-    def test___set____with_shorthand(self) -> None:
-        instance = self._Instance()
-        translation = {
-            DEFAULT_LOCALE_TAG: {
-                "one": "{count} world",
-                "other": "{count} worlds",
-            },
-        }
-        instance.attr = translation  # ty:ignore[invalid-assignment]
-        assert instance.attr.count(2).localize(DEFAULT_LOCALIZER) == "2 worlds"
-
-    def test___set____with_mapping(self) -> None:
-        instance = self._Instance()
-        translation = {
-            DEFAULT_LOCALE: {
-                "one": "{count} world",
-                "other": "{count} worlds",
-            },
-        }
-        instance.attr = translation  # ty:ignore[invalid-assignment]
-        assert instance.attr.count(2).localize(DEFAULT_LOCALIZER) == "2 worlds"
