+++ betty/tests/test_property.py
